Moves logic to display the correct path to AbstractPost instead of URL extension
